--- packages/toolpad-core/src/createQuery.tsx
@@ -1,8 +1,0 @@
-import createFunction from './createFunction.js';
-
-/**
- * @deprecated
- * createQuery is deprecated. Use createFunction instead.
- */
-const createQuery = createFunction;
-export default createQuery;
