+++ packages/toolpad-core/src/index.tsx
@@ -24,13 +24,8 @@ export { default as useUrlQueryState } from './useUrlQueryState.js';
 
 export * from './constants.js';
 
-export { default as createComponent } from './createComponent.js';
-export * from './createComponent.js';
+export * from './browser.js';
 
 export * from './types.js';
 
-export { default as createQuery } from './createQuery.js';
-export * from './createQuery.js';
-
-export { default as createFunction } from './createFunction.js';
-export * from './createFunction.js';
+export * from './server.js';
